fix: Use pipx to install podman-compose on Debian12+
The system Python is marked as externally managed, thus the installation
fails. We should probably migrate to pipx anyways for installation of
Python software in general. In other formulae, I do this via virtualenv,
but it's also a separate package that needs to be installed and pipx is
meant for the job.