fix(compose): Recreate containers when env_file has changes
As of podman-compose v1.2.0, environment variables defined in env_file
contents are passed directly. When they change, the hash does not
display changes, hence the containers are effectively outdated.

This adds another ugly patch on top of the ugly compose modules, which
tries to detect when env_file contents have changed and forces container
recreation.

I really need to migrate to quadlet.